RapidTeleop/requirements.txt
av==14.4.0
avp_stream==1.0
cv_bridge==1.16.2
dex_retargeting==0.5.0
dynamixel_sdk==3.7.31
matplotlib==2.2.3
mediapipe==0.10.21
numpy==1.15.1
pandas==0.23.4
pinocchio==0.4.3
pygame==2.6.1
pyk4a==1.5.0
pynput==1.8.1
pyserial==3.5
pytransform3d==3.14.1
PyYAML==6.0.2
rospy==1.17.0
sapien==2.2.2
scipy==1.1.0
sensor_msgs==1.13.1
setuptools==40.2.0
torch==2.7.1
